tests/requirements.txt
pytest
pytest-asyncio
pytest-timeout
docker
websocket-client
pyserial
behave

